Unified r2o() and r2o2() conversion functions
These functions converted contents of a register to an object. They were
identical except for a different static object in each function. Pointer
to this object was the return value. As the static aspect of the objects
was not used, users of the function now have to provide memory for the
object. Now only r2o() can be used.