Better looping for fetch_answer and update_answer
These functions loop the wait_for_answer queue of the comm_s structure
to find the correct buffer.
Now as the queues are locked the previous solution with dequeuing and
enqueuing was not optimal. New macros for looping the queue with held
mutex were added and if the item is found only one delete operation has
to be performed.
Function comm_buf_del() is used to delete any item from the queue (not
just the first one or the last one). It returns the buffer of the item.