Fix update and fetch race condition
Before this change, an update or a fetch operation could be answered by
the kernel faster than Constable could add the originating buffer to a
queue of waiting buffers.
Now the buffer is added to the queue before it is sent to the kernel.
This change eliminated functions mcp_fetch_object_wait and
mcp_update_object_wait.