Fix buffer resize during read operation
During read operation, the kernel might want to send more data than the
size of the receiving buffer. In that case, the buffer is resized. Since
the pointer to the buffer is stored on stack of another function, we
need to change the pointer to the new buffer.
This commit changes previous comm_buffer_s *buf definitions to
comm_buffer_s **buf.