feat: exit Constable with failure on fetch/update kernel answer
Run `fatal()` on fetch/update kernel answer if there is nobody
waiting for it.

Reason: is an error if arrived an answer to fetch/update request
and there is nothing waiting for it. That means, someone in the
past sent a fetch/update request to the kernel, but for some
reason (an error) the corresponding sender buffer is not enqueued
in the wait queue.