fix: force open() to return fd > 2
Syscall open() returns the first unused fd for a calling process.
Constable communicates with the kernel via /dev/medusa and for
this purpose it uses open() syscall. If Constable is running from
the init process directly from the kernel via UMH, the standard
fds (0, 1 and 2) are not used. So open("/dev/medusa") returns fd
value less than 3. This leads to spurious writes to /dev/medusa
when debug messages via [f]printf() are printed out.

Solution:
Create in the module responsible for the communication (comm.[h|c])
a new function comm_open_skip_stdfds(fname, flags, mode), which
try to open a file @fname. If success, check returned value of the
file descriptor from open() syscall and in the case it's below 3
dup() returned fd until it reaches first value above 2.

From a portability and extension point of view of the Constable
it's not a good idea use the dup2(oldfd, newfd) syscall, because
"if the file descriptor newfd was previously open, it is closed
before being reused; the close is performed silently (i.e., any
errors during the close are not reported by dup2())." [man dup2].
To prevent accidental closing of @newfd comm_open_skip_stdfds()
uses dup() syscall.

Constable opens not only /dev/medusa, but if an option -D[D] or -d
is used, there can be opened another files. So they are also users
of the new comm_open_skip_stdfds() API to not confuse writing to
them with debug messages via [f]printf().