fix: infinite loop in comm_open_skip_stdfds()
original code:
1    while fd < 3:
2        tmp := dup(fd)
3        close(fd)
4        fd := tmp

introduced an infinity loop, because dup() syscall allocates "the
*lowest-numbered* file descriptor that was unused in the calling
process" [man dup]. Syscall close() on line 3 closes the fd, which
value is below 3 (see the condition on line 1), so in the next
iteration dup() returns value below 3: or the value of the fd
closed in previous turn or another below it. This leads to an
infinity loop.

Example for input fd 1
    turn 1: dup(1) return 2, close fd 1
    turn 2: dup(2) return *1*, close fd 2
    turn 3: dup(1) return 2, close fd 1
    ...

Fixes: 4375df6 (fix: force open() to return fd > 2)
Co-authored-by: Roderik Ploszek <roderik.ploszek@gmail.com>