feat: change interface of event_type_find_name()
The semantics of event_type_find_name(name) is:
Find event description by @name in the global list `events' and
returns it.

But if the event description for @name is not present in the
global `events' list, the function allocated and registered new
event description for @name.

There are two use cases of event_type_find_name() in Constable
code base:
1) calls, where a new event description is desired, and
2) calls, where shouldn't be allocated a new event description.

The interface of the function does not allowed distinguishing
between the two use cases, so this commit changes it.