fix: the loss of memory block(s) in space_add_path()
The only user of new_path() is space_add_path(). It calls
new_path() and the result stores into `prev'. To prevent the loss
of memory block(s) in case of an error by rewriting `prev' in the
caller with a NULL value, the entire list @prev is freed before
returning NULL from new_path().