feat(space): get an array of real members of a space
The attribute ->ltree of struct space_s stores mixed information
about space members: a member can be of struct tree_s (i.e. a node
of UNST) or struct space_s (i.e. another space).

Moreover, a member can be stored multiple times in the list and
there can be members which are marked as excluded from the space.

So this commit stores final (real) list of members of the space
into ->members of struct space_s. This is done after evaluation
of space members in space_apply_all(). The final list
->members.array stores deduplicated pointers to UNST nodes (real
members of the space) and their count in ->members.count field.