feat(space): apply func(arg) to each member of a space
Create a function space_for_every_member(space, func, arg), which
applies func(arg) to each member of the space. The function is
a copy of space_for_every_path(), but for members enumerating
uses a space->members struct (where all members of the space are
stored) instead of dynamic members enumeration as in case of
space_for_every_path().

Before this commit an access to space members was done with
dynamic enumeration. Now, only the list of members of a space is
obtained with space_for_every_path() function and all other calls
of space_for_every_path() are replaced with the new function
space_for_every_member().

Benefits of this feature: the direct access to all members of a
space.