fix(constable): runtime() and its users
If defined %DEBUG_TRACE macro, runtime() uses thread specific
variables for some information useful for example in event
handler execution. But these information are not present (or set)
in all users of runtime().

So if %DEBUG_TRACE is set and the additional information are not
present, runtime() prints meaningless output. This commit changes
this behavior. The additional information are printed, only if
all (additional) items are present. Otherwise, the function
behaves as if the %DEBUG_TRACE is **not** set.